requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0
requests>=2.25.0
tqdm>=4.62.0
customtkinter>=5.1.2
tkinter-tooltip>=2.0.0
ttkbootstrap>=1.10.1